+++ src/print_tasks/AbstractPrintTask.ts
@@ -1,4 +1,4 @@
-import { PrintPacketProgressEvent } from "..";
+import { PrintEndEvent, PrintPacketProgressEvent, PrintProgressEvent } from "..";
 import { EncodedImage } from "../image_encoder";
 import { LabelType, PageColorType } from "../packets";
 import { NiimbotProtocol } from "../packets/protocol";
@@ -154,8 +154,10 @@ export abstract class AbstractPrintTask {
     return this.protocol.getClient().getModelMetadata()?.printheadPixels;
   }
   /** End print, cleanup */
-  printEnd(): Promise<boolean> {
-    return this.protocol.printEnd();
+  async printEnd(): Promise<boolean> {
+    const result = await this.protocol.printEnd();
+    this.emitPrintEndEvent();
+    return result;
   }
   /** Check if this print task supports a specified page color */
   isSupportColor(pageColor: PageColorType): boolean {
@@ -171,8 +173,21 @@ export abstract class AbstractPrintTask {
 
       if (step > lastReportedStep || current === total) {
         lastReportedStep = step;
-        this.protocol.getClient().emit("printpacketprogress", new PrintPacketProgressEvent(percent))
+        this.protocol.getClient().emit("printpacketprogress", new PrintPacketProgressEvent(percent));
       }
     };
+  }
+
+  protected emitProgressEvent(pagePrintProgress: number = 100, pageFeedProgress: number = 100) {
+    this.protocol
+      .getClient()
+      .emit(
+        "printprogress",
+        new PrintProgressEvent(this.printOptions.totalPages, this.pagesPrinted, pagePrintProgress, pageFeedProgress),
+      );
+  }
+
+  protected emitPrintEndEvent() {
+    this.protocol.getClient().emit("printend", new PrintEndEvent(this.pagesPrinted));
   }
 }